+++ CHANGELOG.md
@@ -418,6 +418,7 @@
 * [X] 重构和添加`queryBy`、`getBy`、`getOrNullBy`、`getOrElseBy`
 * [X] 进一步实现`PathType`
 * [X] 移动相关扩展方法到`BreezeComponentExtensions`
+* [X] 重命名`LetterCase`为`CaseType`
 * [ ] 整理`breeze-dsl`中
 * [ ] 用于处理CSV和TSV的`Mapper`